requirements.txt
agentops==0.3.4
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.26.4
openai==1.37.1
outcome==1.3.0.post0
packaging==23.2
psutil==5.9.8
pyautogen==0.2.32
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==1.0.1
regex==2024.7.24
requests==2.31.0
selenium==4.23.1
sniffio==1.3.1
sortedcontainers==2.4.0
termcolor==2.4.0
tiktoken==0.7.0
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
